requirement.txt
langchain-openai
langchain
langchain_community
langchain_chroma
tiktoken
chromadb
bs4
torch
torchvision
torchaudio
faiss-cpu
sentence-transformers
transformers
huggingface-hub
faiss-gpu
requests==2.28.1
wget==3.2
langchain_experimental
tavily-python
streamlit
protobuf==3.20.*
sendgrid